reorg and init integration with hybrid consensus
