+++ trueconsensus/fastchain/__init__.py
@@ -2,6 +2,5 @@
 Package import structure
 '''
 
-from fastchain.bft import *
+from fastchain.bft_committee import *
 from fastchain.node import *
-
