+++ trueconsensus/fastchain/local_config.py
@@ -2,11 +2,11 @@ import os
 
 # CFG_YAML_PATH = "/etc/pbft_tunables.yaml"
 # CFG_YAML_PATH = os.path.join(os.path.expanduser('~'), "pbft_tunables.yaml")
-CFG_YAML_PATH = "pbft_tunables.yaml"
+CFG_YAML_PATH = "conf/pbft_tunables.yaml"
 
 # CFG_GENERAL_PATH = "/etc/pbft_logistics.cfg"
 # os.path.join(os.path.expanduser('~'), "pbft_logistics.cfg")
-CFG_GENERAL_PATH = "pbft_logistics.cfg"
+CFG_GENERAL_PATH = "conf/pbft_logistics.cfg"
 
 # list of boot nodes
 PEER_NETWORK_FILE = os.path.join(os.path.expanduser('~'), "hosts")
