--- trueconsensus/fastchain/proto.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-protoc --python_out=./ request.proto
